+++ app/(front)/about/hr.tsx
