+++ app/(front)/about/subtitle.tsx
