+++ app/(front)/announcements/page.tsx
