+++ app/(front)/clubs/actions.ts
@@ -1,6 +1,6 @@
 'use server'
 
-import {createClient} from '@/utils/supabase/server'
+import {createClient} from '@/utils/supabase/server.ts'
 
 export type Club = {
   id: number
