+++ app/(front)/contact/map.tsx
@@ -1,4 +1,4 @@
-import {Card, CardContent} from '@/components/ui/card'
+import {Card, CardContent} from '@/components/ui/card.tsx'
 
 export function Map() {
   return (
