+++ app/(front)/page.tsx
