+++ app/(front)/services-section.tsx
@@ -3,8 +3,8 @@ import {
   faHandHoldingHand,
   faGraduationCap
 } from '@fortawesome/free-solid-svg-icons'
-import {ServiceCard} from '@/components/service-card'
-import {Section} from '@/components/section'
+import {ServiceCard} from '@/components/service-card.tsx'
+import {Section} from '@/components/section.tsx'
 
 export function ServicesSection() {
   return (
