+++ app/(news)/ang-pararayos/layout.tsx
